pyproject.toml
[project]
name = "polybot"
version = "0.9.5"
description = "A simple framework for building robust social media bots for multiple networks"
authors = [{ name = "Russ Garrett", email = "russ@garrett.co.uk" }]
readme = "README.md"
# atproto doesn't currently work on 3.13
requires-python = "<3.13,>=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "atproto>=0.0.55",
    "httpx>=0.27.2",
    "mastodon-py>=1.8.1",
    "pillow>=11.0.0",
    "tweepy>=4.14.0",
]

[project.urls]
Homepage = "https://github.com/russss/polybot"

[dependency-groups]
dev = ["mypy>=1.13.0", "ruff>=0.8.1", "pytest>=8.3.4"]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
line-length = 110

[tool.ruff.lint]
select = ["E", "F", "UP", "I"]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true
